==> 01_MIT_BIH_preprocessing.ipynb <==
from utils import *
from scipy.fft import fft, fftfreq
from matplotlib import pyplot as plt

df_record_lead_ann = pd.read_parquet(join(dataframes_directory, 'df_record_lead_ann.parquet'))
df_lead_ann_summery =  pd.read_parquet(join(dataframes_directory, 'df_lead_ann_summery.parquet'))

list_leads = ['MLII']#, 'V1']
mask_leads = df_record_lead_ann.upper_signal.isin(list_leads)# & df_record_lead_ann.lower_signal.isin(list_leads)
# filter leads
df_record_lead_ann = df_record_lead_ann[mask_leads].reset_index(drop=True)

df_record_lead_ann#.head()

# create figure with subplots
fig, axes = plt.subplots(df_record_lead_ann.shape[0], 1, figsize=(5, 80))

# Tight layout often produces nice results
# but requires the title to be spaced accordingly
fig.tight_layout()


for plot_num, record_num in enumerate(df_record_lead_ann.record):
    ax = axes[plot_num]
    record, ann =  load_record(record_num)
    signal = record.p_signal[:,0]
    N = signal.shape[0]
    fs = record.fs
    T = 1/fs
    yf = 2.0/N *np.abs(fft(signal)[:N//2])
    xf = fftfreq(N, T)[:N//2]
    ax.plot(xf[1:], yf[1:])
    ax.set_xlim(-1, fs/2)
    ax.grid()
    ax.set_title(f'{record_num} {df_record_lead_ann.upper_signal[plot_num]}')
fig.savefig(join(figures_directory, 'MLII_spectrum.svg'), dpi=300)
